Daily Code 10/29/18
Write a text based number-guessing game program. 
The program picks a number between 1 and 100 and asks the user to guess what number it has chosen. If the user guesses wrong, the program tells them whether they're too high or too low. When the user guesses correctly, the program prints out a congratulatory message and tells them how many turns it took them.
When the player finally wins, the program should ask them if they want to play again, and if not, quit the game.